HIPCMS-420 Added TopicPermissions

+++ src/Api/Controllers/TopicsController.cs
@@ -10,6 +10,7 @@ using Api.Data;
 using Microsoft.AspNetCore.Http;
 using System.IO;
 using Api.Models.Entity;
+using Api.Permission;
 
 namespace Api.Controllers
 {
@@ -18,14 +19,16 @@ namespace Api.Controllers
         private TopicManager topicManager;
         private AttachmentsManager attachmentsManager;
 
+        private TopicPermissions topicPermissions;
+
         public TopicsController(CmsDbContext dbContext, ILoggerFactory loggerFactory) : base(dbContext, loggerFactory)
         {
             topicManager = new TopicManager(dbContext);
             attachmentsManager = new AttachmentsManager(dbContext);
+            topicPermissions = new TopicPermissions(dbContext);
         }
 
-        #region topics
-
+        #region GET topics
 
         // GET api/topics
         [HttpGet]
@@ -43,6 +46,9 @@ namespace Api.Controllers
         [HttpGet("{id}")]
         public IActionResult Get(int id)
         {
+            if (!topicPermissions.IsAssociatedTo(User.Identity.GetUserId(), id))
+                return Unauthorized();
+
             var topics = topicManager.GetTopicById(id);
             if (topics != null)
                 return Ok(topics);
@@ -50,35 +56,13 @@ namespace Api.Controllers
             return NotFound();
         }
 
-
-        // GET api/topics/:id/students
-        [HttpGet("{id}/Students")]
-        public IActionResult GetTopicStudents(int id)
-        {
-            return Ok(topicManager.GetAssociatedUsersByRole(id, Role.Student));
-        }
-
-
-        // GET api/topics/:id/supervisors
-        [HttpGet("{id}/Supervisors")]
-        public IActionResult GetTopicSupervisors(int id)
-        {
-            return Ok(topicManager.GetAssociatedUsersByRole(id, Role.Supervisor));
-        }
-
-
-        // GET api/topics/:id/reviewers
-        [HttpGet("{id}/Reviewers")]
-        public IActionResult GetTopicReviewers(int id)
-        {
-            return Ok(topicManager.GetAssociatedUsersByRole(id, Role.Reviewer));
-        }
-
-
         // GET api/topics/:id/subtopics
         [HttpGet("{id}/SubTopics")]
         public IActionResult GetSubTopics(int id)
         {
+            if (!topicPermissions.IsAllowedToEdit(User.Identity.GetUserId(), id))
+                return Unauthorized();
+
             return Ok(topicManager.GetSubTopics(id));
         }
 
@@ -87,15 +71,57 @@ namespace Api.Controllers
         [HttpGet("{id}/ParentTopics")]
         public IActionResult GetParentTopics(int id)
         {
+            if (!topicPermissions.IsAllowedToEdit(User.Identity.GetUserId(), id))
+                return Unauthorized();
+
             return Ok(topicManager.GetParentTopics(id));
         }
+
+        #endregion
+
+        #region Get Users
+
+        // GET api/topics/:id/students
+        [HttpGet("{id}/Students")]
+        public IActionResult GetTopicStudents(int id)
+        {
+            return GetTopicUsers(id, Role.Student);
+        }
+
+        // GET api/topics/:id/supervisors
+        [HttpGet("{id}/Supervisors")]
+        public IActionResult GetTopicSupervisors(int id)
+        {
+            return GetTopicUsers(id, Role.Supervisor);
+        }
+        
+        // GET api/topics/:id/reviewers
+        [HttpGet("{id}/Reviewers")]
+        public IActionResult GetTopicReviewers(int id)
+        {
+             return GetTopicUsers(id, Role.Reviewer);
+        }
+
+        private IActionResult GetTopicUsers(int id, string role)
+        {
+            if (!topicPermissions.IsAllowedToEdit(User.Identity.GetUserId(), id))
+                return Unauthorized();
+
+            return Ok(topicManager.GetAssociatedUsersByRole(id, role));
+        }
+
+        #endregion
+
+        #region Edit Topics
 
 
         // POST api/topics
         [HttpPost]
-        [Authorize(Roles = Role.Supervisor)]
         public IActionResult Post(TopicFormModel model)
         {
+            if (!topicPermissions.IsAllowedToAdminister(User.Identity.GetUserId()))
+                return Unauthorized();
+
             if (ModelState.IsValid)
             {
                 if (!Status.IsStatusValid(model.Status))
@@ -118,10 +144,13 @@ namespace Api.Controllers
         [Authorize(Roles = Role.Supervisor)]
         public IActionResult Put(int id, TopicFormModel model)
         {
+            if (!topicPermissions.IsAllowedToEdit(User.Identity.GetUserId(), id))
+                return Unauthorized();
+
             if (ModelState.IsValid)
             { // TODO createUser is Supervisor!
                 if (topicManager.UpdateTopic(User.Identity.GetUserId(), id, model))
-                    return Ok();
+                    return Ok(42);
             }
             return BadRequest(ModelState);
         }
@@ -130,6 +159,9 @@ namespace Api.Controllers
         [HttpPut("{id}/Status")]
         public IActionResult ChangeStatus(int id, string status)
         {
+            if (!topicPermissions.IsAllowedToEdit(User.Identity.GetUserId(),id))
+                return Unauthorized();
+
             if (!Status.IsStatusValid(status))
                 ModelState.AddModelError("status", "Invalid Status");
             else if (topicManager.ChangeTopicStatus(User.Identity.GetUserId(), id, status))
@@ -141,9 +173,10 @@ namespace Api.Controllers
 
         // DELETE api/topics/:id
         [HttpDelete("{id}")]
-        [Authorize(Roles = Role.Supervisor)]
         public IActionResult Delete(int id)
         {
+            if (!topicPermissions.IsAllowedToAdminister(User.Identity.GetUserId()))
+                return Unauthorized();
             if (topicManager.DeleteTopic(id))
                 return Ok();
 
@@ -159,6 +192,9 @@ namespace Api.Controllers
         [HttpGet("{topicId}/Attachments")]
         public IActionResult GetAttachments(int topicId)
         {
+            if (!topicPermissions.IsAssociatedTo(User.Identity.GetUserId(), topicId))
+                return Unauthorized();
+
             var attachments = attachmentsManager.GetAttachments(topicId);
             if (attachments != null)
                 return Ok(attachments);
@@ -169,6 +205,9 @@ namespace Api.Controllers
         [HttpGet("{topicId}/Attachments/{attachmentId}")]
         public IActionResult GetPicture(int topicId, int attachmentId)
         {
+            if (!topicPermissions.IsAssociatedTo(User.Identity.GetUserId(), topicId))
+                return Unauthorized();
+
             var attachment = attachmentsManager.GetAttachmentById(topicId, attachmentId);
             if (attachment != null)
             {
@@ -182,6 +221,9 @@ namespace Api.Controllers
         [HttpPost("{topicId}/Attachments")]
         public IActionResult PostAttachment(int topicId, AttatchmentFormModel model, IFormFile file)
         {
+            if (!topicPermissions.IsAllowedToEdit(User.Identity.GetUserId(), topicId))
+                return Unauthorized();
+
             if (file == null)
                 ModelState.AddModelError("file", "File is null");
 
@@ -200,6 +242,9 @@ namespace Api.Controllers
         [HttpDelete("{topicId}/Attachments/{attachmentId}")]
         public IActionResult DeleteAttachment(int topicId, int attachmentId)
         {
+            if (!topicPermissions.IsAllowedToEdit(User.Identity.GetUserId(), topicId))
+                return Unauthorized();
+
             if (attachmentsManager.DeleteAttachment(topicId, attachmentId))
                 return Ok();
             return BadRequest();
